move cuda out of core.cc; test wK
