+++ psi4/src/CMakeLists.txt
@@ -66,6 +66,16 @@ target_sources(core
     python_data_type.cc
   )
 
+# GPU discovery/health checking and cuEST+CUDA handle lifetime, kept out of
+# core.cc. Only compiled when cuEST is enabled; core.cc includes
+# cuest_runtime.h unconditionally and gets no-op inline stubs otherwise.
+if(TARGET cuEST::cuEST)
+  target_sources(core
+    PRIVATE
+      cuest_runtime.cc
+    )
+endif()
+
 get_property(psi4_binmodules GLOBAL PROPERTY LINKBINLIST)
 get_property(psi4_libmodules GLOBAL PROPERTY LINKLIBLIST)
 target_link_libraries(core
